Controllers, Facades, Repositories,Services, DTO for Comment, Like, Post, Subscription GlobalExceprionHandler

+++ src/main/java/com/example/social_network_backend/Services/UserService.java
@@ -5,11 +5,12 @@ import com.example.social_network_backend.DTO.User.UpdateUserDTO;
 import com.example.social_network_backend.Entities.User;
 import com.example.social_network_backend.Repositories.UserRepository;
 import jakarta.persistence.EntityNotFoundException;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import lombok.AllArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
+import org.springframework.transaction.annotation.Transactional;
 
 @AllArgsConstructor
 @Service
@@ -18,20 +19,23 @@ public class UserService {
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
 
-
+    @Transactional
     public User createUser(CreateUserDTO dto) {
-        return mapToEntity(dto);
+        User user = mapToEntity(dto);
+        user.setPassword(passwordEncoder.encode(user.getPassword())); // Hash pass
+        userRepository.save(user);
+        return user;
     }
 
     public User getUserById(Long id) {
         return userRepository.findById(id).orElseThrow(EntityNotFoundException::new);
     }
 
-    public List<User> getAllUsers() {
-        User user = new User();
-        return List.of(user);
+    public Page<User> getAllUsers(Pageable pageable) {
+        return userRepository.findAll(pageable);
     }
 
+    @Transactional
     public User updateUser(Long id, UpdateUserDTO dto) {
         User user = userRepository.findById(id).orElseThrow(EntityNotFoundException::new);
         user.setName(dto.name());
@@ -41,15 +45,19 @@ public class UserService {
         return user;
     }
 
+    @Transactional
     public void deleteUser(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new EntityNotFoundException();
+        }
         userRepository.deleteById(id);
     }
 
     public String authenticate(String email, String password) {
-        String testEmail = "john.doe@example.com";
-        String testPassword = passwordEncoder.encode("password123");
-        if (email.equals(testEmail) && passwordEncoder.matches(password, testPassword)) {
-            return "dummy-jwt-token"; // Временный токен
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new EntityNotFoundException("User not found"));
+        if (passwordEncoder.matches(password, user.getPassword())) {
+            return "dummy-jwt-token"; // JWT
         } else {
             throw new RuntimeException("Invalid credentials");
         }
@@ -60,8 +68,9 @@ public class UserService {
         user.setName(dto.name());
         user.setSurname(dto.surname());
         user.setEmail(dto.email());
-        user.setPassword(dto.password());
+        user.setPassword(passwordEncoder.encode(dto.password())); // Хешируем пароль
         user.setSex(dto.sex());
+        user.setRole(dto.role());
         return user;
     }
 }
